refactor: organize restore session tests by contract
